Update the request-executor.jsp.
Add an input textarea to write sql requests, a form to send sql requests to a server.

Add a description of user capabilities depending on the role.

Add new locales for the request-executor.jsp.

Add the getRequestDTO() method in the RequestExecutorController class. It is necessary to automatically add DTO to the page.

Update the RequestDTOValidatorImpl class. Now the validator does not perform additional checks if errors were previously detected.

Write several test (QueryHistoryControllerTest, WelcomeControllerTest) Add a new method in the RequestExecutorControllerTest class.