Finish the full basic version of the project. Version 1.0.
Finish functionality to display the results of the query on the page.

Finish functionality to display the history of requests on the page.

Add a new dependency in pom.xml used in controllers and classes of the DAO layer. The changes were made in them.

Add new beans to context files and update old ones.

Update tests in accordance with the changed behavior of the tested classes.

Add new jsp pages and update old ones.

Add new locale properties.

Add new classes.